refactor: reorganise command processing
The code in the `main` module of the node manager was becoming unwieldy and difficult to navigate,
so I've refactored it into a few different modules.

There are absolutely no functional changes here.

Clippy was complaining about some functions having too many arguments, so I suppressed those
warnings. These are just utility functions, and adding new structs just to encapsulate a few
arguments would have been tedious and overkill.
